Fixed coloring of login stuff, and reorganized some folder structure
